backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
pypdf==3.15.1
openai==1.1.1
pydantic==2.4.2
python-multipart==0.0.6
